fix: 560 mypy-Fehler behoben (Typ-Reparaturen, Laufzeit-Bugs, Vendor-Override)
aurik-mypy-real-bug-gate: 560 -> 0 Fehler (verifiziert: Gate + Unit-Tests + Laufzeit).

- no-any-return (388): cast() an Third-Party-/ONNX-/torch-Grenzen
  (Laufzeit-No-op), Layout-/Mehrzeilen-Sicherheit geprüft
- var-annotated (45) + arg-type/assignment: float()/int()-Konversionen
  (wertidentisch) bzw. getattr/librosa-Stub-Ignores
- pyproject.toml: plugins._vendor_aero von mypy ausgenommen
  (Vendor-Policy, unverändert kopiert — AGENTS.md-konform)

Dabei gefundene echte Laufzeit-Bugs repariert (Gate-Erkenntnisgewinn):
- defect_scanner.is_audible: NameError bei jedem Aufruf (undefinierte
  Band-/PANNs-Namen) -> parameterbasierte Hörbarkeitsprüfung
- quality_gate.check_ml/check_gui: fehlendes return True im Happy-Path +
  doppeltes except (nur der erste griff) -> bool statt None
- silence_mask.apply_silence_preservation: doppeltes except -> Fallback
  lieferte None statt Array
- ast_plugin: (labels,scores)-Unpack auf AstResult-Objekt -> TypeError;
  discriminate_defect mit falscher Signatur -> immer 0.0
- lyrics_transcriber: _decode_result existiert nicht -> DSP-Fallback;
  miipher_dit: register_plugin als Methode statt Modulfunktion
- phase_06: _hann_crossfade nie definiert -> NameError
- phase_42: PhaseResult mit ungültigen kwargs; bridge.inject_cd_noise_profile:
  Signatur passte nicht zu core -> CD-Profil lief nie
- unified_restorer_v3: self.config.sample_rate existiert nicht ->
  §G90-Blöcke liefen nie (jetzt sample_rate)
- sota_vocal_pipeline: tuple/dict-Verwechslung; vocoder_chain: synthesize
  -> reconstruct; evaluation_system: MOSResult.mos; genre_denoise_router:
  fehlendes top3-Argument

Verifikation: Gate 0 Fehler, 303 Unit-Tests grün, Laufzeit-Checks der
reparierten Funktionen grün, ruff F821/F601/B009/I001 clean, py_compile clean.

Bekannt verbleibend (vorbestehend, dokumentiert): der Plain-mypy-Hook
(mit Import-Verfolgung) meldet 23 Fehler — Redundant-Cast-Modus-Konflikt
(skip-imports-Gate vs. with-imports-Hook) plus vorbestehende Fehler im
Import-Graph (denker/, models/). Nicht Teil dieses Fixes.

+++ backend/core/do_no_harm_guardian.py
@@ -371,7 +371,7 @@ class DoNoHarmGuardian:
             _spec_db -= np.max(_spec_db)
             _spec_lin = 10.0 ** (_spec_db / 20.0)
             _spec_norm = _spec_lin / (np.sum(_spec_lin) + 1e-10)
-            _entropy = -np.sum(_spec_norm * np.log2(_spec_norm + 1e-10))
+            _entropy = float(-np.sum(_spec_norm * np.log2(_spec_norm + 1e-10)))
             _max_entropy = np.log2(len(_spec_norm))
             naturalness = float(np.clip(_entropy / max(_max_entropy, 1.0), 0.0, 1.0))
         except Exception:
